# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.User:
    properties:
      create_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_available:
        type: boolean
      is_visible:
        type: boolean
      name:
        type: string
      password:
        type: string
      role_id:
        type: string
      sername:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - role_id
    - sername
    - username
    type: object
  users.UserIn:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      sername:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - sername
    - username
    type: object
  users.UserOut:
    properties:
      create_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      sername:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - name
    - role
    - sername
    - username
    type: object
...
  utils.HttpResponse-users_UserOut:
    properties:
      code_error:
        type: integer
      data:
        $ref: '#/definitions/users.UserOut'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service d'authentification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CMagic Auth
  version: 1.0.0
paths:
...
  /users:
    get:
      description: Récupère la liste de tous les utilisateurs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      summary: Récupère tous les utilisateurs
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Crée un nouvel utilisateur avec les détails fournis.
      parameters:
      - description: Body data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UserIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.HttpResponse-users_UserOut'
      summary: Crée un nouvel utilisateur
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Met à jour un utilisateur en fonction de son ID avec les détails
        fournis.
      parameters:
      - description: Détails de l'utilisateur à mettre à jour
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UserIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpResponse-users_UserOut'
      summary: Met à jour un utilisateur
      tags:
      - Users
  /users/{id}:
    delete:
      description: Supprime un utilisateur en fonction de son ID.
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Aucun contenu
      summary: Supprime un utilisateur
      tags:
      - Users
    get:
      description: Récupère un utilisateur en fonction de son ID.
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Récupère un utilisateur par ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Met à jour un utilisateur en fonction de son ID avec les détails
        fournis.
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: string
      - description: Body data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UserIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpResponse-users_UserOut'
      summary: Met à jour un utilisateur
      tags:
      - Users
  /users/{id}/assign-role/{role}:
    post:
      description: Assigner un rôle spécifié à un utilisateur en fonction de son ID.
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      - description: Rôle à assigner
        in: path
        name: role
        required: true
        type: string
      responses:
        "204":
          description: Aucun contenu
      summary: Assigner un rôle à un utilisateur
      tags:
      - Users
  /users/{id}/remove-role/{role}:
    post:
      description: Supprimer un rôle spécifié d'un utilisateur en fonction de son
        ID.
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      - description: Rôle à supprimer
        in: path
        name: role
        required: true
        type: string
      responses:
        "204":
          description: Aucun contenu
      summary: Supprimer un rôle d'un utilisateur
      tags:
      - Users
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"